+++ lib/ts/recipe/emailpassword/api/emailExists.ts
@@ -33,7 +33,7 @@ export default async function emailExists(apiImplementation: APIInterface, optio
         });
     }
 
-    let result = await apiImplementation.emailExistsGET({ email, options });
+    let result = await apiImplementation.emailExistsGET({ email, options, userContext: {} });
 
     send200Response(options.res, result);
     return true;