+++ lib/ts/recipe/thirdparty/utils.ts
@@ -74,7 +74,7 @@ export function findRightProvider(
         }
 
         // otherwise, we return a provider that matches based on client ID as well.
-        return p.get(undefined, undefined).getClientId() === clientId;
+        return p.get(undefined, undefined, {}).getClientId({}) === clientId;
     });
 }
 
@@ -140,28 +140,38 @@ function validateAndNormaliseEmailVerificationConfig(
         createAndSendCustomEmail:
             config?.emailVerificationFeature?.createAndSendCustomEmail === undefined
                 ? undefined
-                : async (user, link) => {
-                      let userInfo = await recipeInstance.recipeInterfaceImpl.getUserById({ userId: user.id });
+                : async (user, link, userContext) => {
+                      let userInfo = await recipeInstance.recipeInterfaceImpl.getUserById({
+                          userId: user.id,
+                          userContext,
+                      });
                       if (
                           userInfo === undefined ||
                           config?.emailVerificationFeature?.createAndSendCustomEmail === undefined
                       ) {
                           throw new Error("Unknown User ID provided");
                       }
-                      return await config.emailVerificationFeature.createAndSendCustomEmail(userInfo, link);
+                      return await config.emailVerificationFeature.createAndSendCustomEmail(
+                          userInfo,
+                          link,
+                          userContext
+                      );
                   },
         getEmailVerificationURL:
             config?.emailVerificationFeature?.getEmailVerificationURL === undefined
                 ? undefined
-                : async (user) => {
-                      let userInfo = await recipeInstance.recipeInterfaceImpl.getUserById({ userId: user.id });
+                : async (user, userContext) => {
+                      let userInfo = await recipeInstance.recipeInterfaceImpl.getUserById({
+                          userId: user.id,
+                          userContext,
+                      });
                       if (
                           userInfo === undefined ||
                           config?.emailVerificationFeature?.getEmailVerificationURL === undefined
                       ) {
                           throw new Error("Unknown User ID provided");
                       }
-                      return await config.emailVerificationFeature.getEmailVerificationURL(userInfo);
+                      return await config.emailVerificationFeature.getEmailVerificationURL(userInfo, userContext);
                   },
     };
 }